requirements.txt
numpy
pillow
flask
torchvision==0.8.1
torch===1.7.1
